Add basic linting checks
Just add a script ro run gofmt and golint on the source files under cmd
and controller.  This is much cleaner if we break things up a little
bit, so in the interest of making things easier, and allowing some
flexbility for potential future growth, we also move the main
provisioner pkg under a `cmd` directory and add a hack directory for
scripts like `run-lint-checks.sh`.