Implement CSI clone
Implemented support for CSI clone. The 'clone' is a simple
cp command which uses reflink if on an XFS filesystem, other
wise it falls back to a regular copy.

Verify that the source is on the same node as the target otherwise
return an error to the caller. This ensures the pod is scheduled on
the proper node.

Signed-off-by: Alexander Wels <awels@redhat.com>